Refactor AnalyzerService into separate modules
Moved types from `analyzer_service.rs` to `analyzer_service/types.rs` and the service logic to `analyzer_service.rs` (which now acts as the module root). This improves code organization and readability.

- Extracted `ItemKey`, `SaleSummary`, `SaleHistory`, `CheapestListingValue`, `CheapestListings`, `AnalyzerState`, `SoldAmount`, `SoldWithin`, `ResaleStats`, `ResaleOptions` to `types.rs`.
- Updated imports and visibility.
- Preserved existing tests.
